:recycle:(common): centralize message contracts into @trading-model/common
Merge the duplicated message contract types (message, MessageMetadata, IdentifyType,
RoutingType, DeliveryType, SecurityType, BrokerConfig) into a single source of truth
under @trading-model/common/contracts/message.types.

Previously, these types were defined in two locations:
- packages/broker-message/src/shared/types/message.ts
- services/message-manager/src/messaging/core/message.ts
- services/message-manager/src/messaging/broker.type.ts

Both broker-message and message-manager now import from the canonical location.
The old files remain as re-exports for backward compatibility.

Closes #81
